--- a/src/org/mozilla/classfile/ClassFileWriter.java
+++ b/src/org/mozilla/classfile/ClassFileWriter.java
@@ -1598,27 +1598,25 @@ public class ClassFileWriter {
             // locals as the killed block's locals. Ignore uninitialized
             // handlers, because they will also be killed and removed from the
             // exception table.
+            int sbStart = sb.getStart();
             for (int i = 0; i < itsExceptionTableTop; i++) {
                 ExceptionTableEntry ete = itsExceptionTable[i];
                 int eteStart = getLabelPC(ete.itsStartLabel);
-                // we can sort out exception table entries immediately,
-                // that are before or behind current sb
-                if (sb.getEnd() <= eteStart || sb.getStart() >= getLabelPC(ete.itsEndLabel)) {
-                    continue;
-                }
-                if (sb.getStart() > eteStart) {
+                // this is "hot" code and it has been optimized so that 
+                // there are not too many function calls
+                if (sbStart > eteStart && sbStart < getLabelPC(ete.itsEndLabel)) {
                     int handlerPC = getLabelPC(ete.itsHandlerLabel);
                     SuperBlock handlerSB = getSuperBlockFromOffset(handlerPC);
                     locals = handlerSB.getLocals();
                     break;
                 }
-                if (eteStart < sb.getEnd()) {
+                if (eteStart > sbStart && eteStart < sb.getEnd()) {
                     int handlerPC = getLabelPC(ete.itsHandlerLabel);
                     SuperBlock handlerSB = getSuperBlockFromOffset(handlerPC);
                     if (handlerSB.isInitialized()) {
                         locals = handlerSB.getLocals();
+                        break;
                     }
-                    break;
                 }
             }