Hello, I've tried to improve the compile speed, especially, if many exception handlers are involved (see `CodegenTest.testManyExceptionHandlers`)

This test takes in original 33s on my machine.

- step 1: [Use binary search for getSuperBlockFromOffset](https://github.com/mozilla/rhino/commit/5ebfa5dc8456553174ccfbb935ee8ae9a61e2857)
As long as I see, the `itsSuperBlocksStarts` array contains the starts in sorted order, so I use binary search in getSuperblockFromOffset. 

This brings a performance boost down to 3.5s

- step 2: [narrowed lookup range for big exception tables](https://github.com/mozilla/rhino/commit/f0831e048bd9f23e6d51c18f10e5dec2383dc17f)
I also narrowed the lookup range for big exception tables and throw away the exception table entries, that are definitively irrelevant. See https://github.com/mozilla/rhino/commit/f0831e048bd9f23e6d51c18f10e5dec2383dc17f
Now the test runs in 1.2s

- step 3: [Optimized finding locals](https://github.com/mozilla/rhino/commit/73537d69ac15ec40cdbfc704ff8dbea9bb57e059)
A further optimization in finding the locals brings down the test case to 0.6s

- step 4: [TypeInfo.merge do simplest check first](https://github.com/mozilla/rhino/commit/1a7efb599b2dcd6bdd253087de90c9e55682f63b)
This is not really measurable, but showed up as "hot" method in the profiler, so I tried to optimize it also

The "all test time" goes down from about 11:50 to 11:20