@gbrail 2021-09-23
This all looks very good to me. I was hoping it'd speed up test execution for Rhino itself, but sadly it didn't really make a difference. Do you or does anyone else have an example of some complex JavaScript that now compiles a bunch faster so that we can document that for the record?