@rPraml 2021-09-24
@gbrail I've investigated much time in optimizing test execution (compile and run speed) and did only find a few places where optimization is done. (see other PRs) Unfortunately, most of them is not really measurable.
I've only found the `CodegenTest.testManyExceptionHandlers` test, where this change makes a huge difference:

As far as I see, the amount of exception handlers goes into compile speed with a complexity of ~ `O(n^3)` and the CodegenTest creates 1000 of them. If you double the number, the test takes ~8 times longer.

There were 3 nested for - loops, one of them could be replaced with the binary search, so that you'll get `O(n^2 log(n))`

All in all, compile speed will only increase if you have many of exception blocks (which may be an edge case)

BTW: You can increase the testexecution speed by factor ~2 if jacoco is disabled.