@tuchida 2021-09-24
- #915

Rhino is not currently testing the try statement in test262.
I added `language/statements/try` to `test262.properties`, and found that the number of errors was the same before and after the modification.
```
language/statements/try 117/195 (60.0%)
```